refactor : 클라이언트 파일 구조 변화
from
/client/client.c

to
/client/include/client.h
/client/include/connection.h
/client/include/signal_handlers.h
/client/src/client.c
/client/src/connection.c
/client/src/signal_handlers.c
